finished app mainly. Last completed things: loans full, services screen, identification kyc full(submittion error 500 code from server). monitoring full. payments full. transfers full. fix bugs

+++ entity/src/main/java/uz/mobiler/gita/entity/source/remote/response/LoanResponse.kt
@@ -2,9 +2,9 @@ package uz.mobiler.gita.entity.source.remote.response
 
 data class LoanResponse (
     val id:String,
-    val totalAmount: Long,
-    val remaining:Long,
-    val monthlyPayment:Long,
+    val totalAmount: Double,
+    val remaining:Double,
+    val monthlyPayment:Double,
     val termMonths:Long,
     val nextDueDate:String,
     val status: String
